Fix "aqua: not found" in container example
Since the installer is run as `root`, `aqua` gets installed to `root`'s `$HOME`:
```
/root/.local/share/aquaproj-aqua/bin/aqua
```
However, since we add `aqua` to `PATH` after we've switched to the `foo` user, we add `foo`'s `$HOME` to the `PATH`, resulting in `aqua` not being found:
```
~/workspace $ echo $PATH
/home/foo/.local/share/aquaproj-aqua/bin:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin
```
This can either be fixed by moving the installation to after we switch to `foo`, or moving the `PATH` modification to the `root` user. Since `aqua` doesn't require `root`, I've opted to accentuate this by moving the installation to the non-root user.